# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CombatActionResult": {
        "additionalProperties": true,
        "description": "11 个动作类端点共用：/action / /attack-roll / /damage-roll / /spell /\n/spell-roll / /spell-confirm / /smite / /grapple-shove / /class-feature /\n/maneuver / /reaction",
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "attacker_hp_current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attacker Hp Current"
          },
          "combat_over": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Combat Over"
          },
          "damage": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Damage"
          },
          "dice_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dice Result"
          },
          "hit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit"
          },
          "is_crit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Crit"
          },
          "log_msg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Msg"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "target_hp_current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Hp Current"
          }
        },
        "title": "CombatActionResult",
        "type": "object"
      },
      "CombatStateResponse": {
        "additionalProperties": true,
        "description": "对应 api.combat.info.get_combat_state 的返回。",
        "properties": {
          "current_turn_index": {
            "default": 0,
            "title": "Current Turn Index",
            "type": "integer"
          },
          "entities": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EntitySnapshot"
            },
            "default": {},
            "title": "Entities",
            "type": "object"
          },
          "entity_positions": {
            "additionalProperties": true,
            "default": {},
            "title": "Entity Positions",
            "type": "object"
          },
          "grid_data": {
            "additionalProperties": true,
            "default": {},
            "title": "Grid Data",
            "type": "object"
          },
          "round_number": {
            "default": 1,
            "title": "Round Number",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "turn_order": {
            "default": [],
            "items": {},
            "title": "Turn Order",
            "type": "array"
          },
          "turn_states": {
            "additionalProperties": true,
            "default": {},
            "title": "Turn States",
            "type": "object"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "CombatStateResponse",
        "type": "object"
      },
...
      "EntitySnapshot": {
        "additionalProperties": true,
        "description": "战斗地图上的实体（玩家 / 队友 / 敌人共用）。",
        "properties": {
          "ac": {
            "title": "Ac",
            "type": "integer"
          },
          "conditions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Conditions",
            "type": "array"
          },
          "derived": {
            "additionalProperties": true,
            "default": {},
            "title": "Derived",
            "type": "object"
          },
          "hp_current": {
            "title": "Hp Current",
            "type": "integer"
          },
          "hp_max": {
            "title": "Hp Max",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_enemy": {
            "default": false,
            "title": "Is Enemy",
            "type": "boolean"
          },
          "is_player": {
            "default": false,
            "title": "Is Player",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "hp_current",
          "hp_max",
          "ac"
        ],
        "title": "EntitySnapshot",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReactionRequest": {
        "properties": {
          "reaction_type": {
            "title": "Reaction Type",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          }
        },
        "required": [
          "reaction_type"
        ],
        "title": "ReactionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "DnD 5e AI跑团平台后端",
    "title": "AI TRPG Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/game/combat/{session_id}": {
      "get": {
        "description": "获取当前战斗状态（含完整实体数据）",
        "operationId": "get_combat_state_game_combat__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombatStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Combat State",
        "tags": [
          "combat"
        ]
      }
    },
...
    "/game/combat/{session_id}/reaction": {
      "post": {
        "description": "Player uses reaction during enemy turn.\nreaction_type: \"shield\" | \"uncanny_dodge\" | \"hellish_rebuke\" | \"opportunity_attack\"\nCalled by frontend when enemy attacks player and player has reaction available.",
        "operationId": "use_reaction_game_combat__session_id__reaction_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombatActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Use Reaction",
        "tags": [
          "combat"
        ]
      }
    },
...
  }
}